Initial commit of complete game
A truth or dare / spin the bottle style game.

Set the number of players, up to 8; positions the players around the Micro:bit as shown; select the power of the spin; the marker spins round and round until it randomly stops on a selected player; press a button to spin again.

Doesn't have to be used for Truth or Dare!